+++ Source/ACE.DatLoader.Tests/DatTests.cs
@@ -52,7 +52,7 @@ namespace ACE.DatLoader.Tests
                 if (kvp.Value.FileSize == 0) // DatFileType.LandBlock files can be empty
                     continue;
 
-                var fileType = kvp.Value.GetFileType();
+                var fileType = kvp.Value.GetFileType(DatDatabaseType.Cell);
 
                 if ((kvp.Key & 0xFFFF) == 0xFFFE) fileType = DatFileType.LandBlockInfo;
                 if ((kvp.Key & 0xFFFF) == 0xFFFF) fileType = DatFileType.LandBlock;
@@ -107,7 +107,7 @@ namespace ACE.DatLoader.Tests
                 if (kvp.Key == 0xFFFF0001) // Not sure what this is, EOF record maybe?
                     continue;
 
-                var fileType = kvp.Value.GetFileType();
+                var fileType = kvp.Value.GetFileType(DatDatabaseType.Portal);
 
                 Assert.IsNotNull(fileType, $"Key: 0x{kvp.Key:X8}, ObjectID: 0x{kvp.Value.ObjectId:X8}, FileSize: {kvp.Value.FileSize}, BitFlags:, 0x{kvp.Value.BitFlags:X8}");